fix StructuredTopology.periodic with !isdim axes
The property StructuredTopology.periodic returns a list of indices of periodic
axes, where the index incorrectly refers to the list of all axes instead of
only the axes with dimension.  This causes a problem when creating a basis on
the boundary of a domain with a periodic dimension:

    domain, geom = mesh.rectilinear([2,2], periodic=[1])
    domain.boundary['left'].periodic # should return [0]
    len(domain.boundary['left'].basis('spline', degree=1)) # should return 2

Fixed in this commit.