drop support for partially evaluable bases
In the past a `HierarchicalTopology` basis was a masked concatenation of the
bases at all hierarchical levels.  To make this work, basis functions defined
at refined elements with respect to the elements in the hierarchical topology
evaluate to an empty array.  With the new `Polyval`-based basis and the
concatenate-less hierarchical basis in place, this behavior is reverted.