fix parsing $ as eye in expressions
Due to a bad design choice in the expression tokenizer — insert an eye if we
match a variable with a name equal to one of the eye symbols — `$`-symbols were
not tokenized as eyes.  This patch changes the logic: if we match an eye symbol
and at the same position we can't match a (strict) longer variable, we insert
an eye, otherwise the variable.  The same logic is applied to normals.

Fixes #286.