restyle viewer header using css flexbox
The old header used a nested grid for showing elements specific to the mode of
the viewer: log or theater.  As a side effect the width of the label of the log
or theater could not be constrained.  This patch uses a single flexbox and adds
`show-if-theater` and `show-if-log` classes to specific elements.  In the
stylesheet the elements are hidden appropriately.