close cache file before yielding item
When iterating over a `cache.Recursion` the cache file is kept open (and
locked) until the next item is requested (`next(recursion_iter)`), or the
iterator is garbage collected.  The latter leads to undeterministic behavior in
Pypy, and a deadlock in a unittest.  This commit fixes the problem by closing
the file before yielding.