fix importing dynamic treelog attributes
In Python 3.7 treelog has dynamic module functions pointing to the current log
instance.  For example `treelog.info` resolves to `treelog.current.info`, where
`treelog.current` is obtained when `info` is obtained from `treelog`.  In the
current code we import `info` and others from `treelog` in `nutils.log`,

    from treelog import info

which translates to

    import treelog
    info = treelog.current.info

If `treelog.current` is changed after importing `treelog`, this change is not
reflected by the local `info`.

This patch fixes the issue by resolving the attributes dynamically with the
`__getattr__` module function.