don't pop the head of transform chains
In the past transform chains had a special root transform item. This concept
does not exist anymore and we can safely remove `function.PopHead`.
`function.PopHead` was only used in `ApplyTransforms(PopHead(ndims,
SelectChain())`, where `SelectChain` is implicit. Since `function.SelectChain`
does not yet know the dimension of the root, we temporarily replace `PopHead`
with `TransformChainWithTodims`, with the sole purpose of defining the
`todims`. In a follow-up commit `function.SelectChain` will gain this knowledge
and `TransformChainWithTodims` will be removed in its entirety.