add Sample.getindex(ielem)
`Sample.index` is often a tuple of consecutive slices of
`numpy.arange(Sample.npoints)`. If this is the case, it is a waist of memory to
store this attribute, in particular if this attribute is becoming unused for
special-cased implementations of `Sample` for uniform points sequences or
products of samples. This patch adds the `Sample.getindex` method as a lazy
alternative to `Sample.index[ielem]` and replaces occurences of the latter with
the former in the library when possible.