add Topology.{f_index,f_coords} properties
Currently all bases compute the element index and coordinates internally given
a transform chain sequence (and the fromdims of the sequence). With tensorial
topologies the fromdims attribute will be removed, because transform chains in
a sequence may have variable fromdims, and the fromdims should be passed
explicitly to the bases. In addition the tensorial roots should be passed to
the bases as well.

An alternative to the above suggested additional arguments is to replace the
`transforms` argument (and `trans`, necessary for opposites) with `index` and
`coords` (which supports opposites automatically).

This patch adds the convenience properties `f_index` and `f_coords` to
`Topology`. The former is an evaluable element index, the latter the evaluable
element local coordinates. These properties are used in a follow-up patch the
arguments of the bases are changed to `index` and `coords`.