repl. basis args transfs, trans with index, coords
Currently all bases compute the element index and coordinates internally given
a transform chain sequence (and the fromdims of the sequence). With tensorial
topologies the fromdims attribute will be removed, because transform chains in
a sequence may have variable fromdims, and the fromdims should be passed
explicitly to the bases. In addition the tensorial roots should be passed to
the bases as well.

An alternative to the above suggested additional arguments is to replace the
`transforms` argument (and `trans`, necessary for opposites) with `index` and
`coords` (which supports opposites automatically).

This patch replaces the arguments `transforms` and `trans` of bases with
`index` and `coords` and removes the `transforms` attribute.