use summation convention for func calls
Currently the expression syntax supports function calls with multiple arguments
where each argument has the same shape. In a follow-up commit the function
syntax will be used for derivatives: `d(u_i, x_j)` and `d(v, x_i, x_i)` as
alternatives for `u_i,j` and `v_,ii`. This commit applies summation of repeated
indices to function calls, which is necessary for the derivatives. Since
`arctan2` is the only function available in the namespace that takes more than
one argument, the impact of this change is assumed to be negligible.