support grad to nd geom
Given a product topology of space and time, it is more convenient to write a
derivative to time as `d(f, t)` (expression syntax) than as `d(f, t_0)`, where
`t` is the 0D time or the 1D time with shape `(1,)` respectively. This patch
adds support for the former and for completeness for an nD geometry.