improve implementation of Kronecker._get
The current frozenarray implementation tests for equality based on contents and
dtype, meaning that an int32 and int64 array will never test equal. This could
lead to an error in Kronecker._get, which used to test for equality of two
arrays (valid for Numpy arrays of size 1) without realizing that these arrays
were, or could be, frozen. The new implementation unpacks the scalar values
prior to comparison which takes care of this particular bug. A followup commit
will fix frozenarray's overall behaviour.