remove basis from namespace pickle test
In a follow-up commit the `function` module will be rebuild from the ground up,
Since this module is nothing more than a thin wrapper around the `evaluable`
module, which incorporates automatic deduplication, it is unnecessary to add
deduplication to the classes in the rebuild `function` module as well. This has
one important side-effect. The bases, which now become part of the `function`
module only, use `evaluable.ElemwiseFromCallable` inside `prepare_eval` to
generate an evaluable, referring to methods of the bases. This unfortunately
breaks the deduplication of the `ElemwiseFromCallable`.

The `namespace.test_pickle` test relies on deduplication after unpickling,
which, in case a basis is pickled, will fail in the future. This patch sweeps
the problem under the carpet by simply not pickling a basis at all.