remove points axis from coeffs in poly_eval
Currently `numeric.poly_eval` assumes the coefficients and the coordinates have
exactly one leading points axis. This patch prepares `numeric.poly_eval` for a
variable number of leading points axes in the future.

The most flexible extension is to force both the coefficients and the
coordinates to have the same leading axes. The coordinates gets one additional
axis of length `ndim` and the cooefficients `ndim` additional axes. This,
however, has a detrimental effect on the performance of `numeric.poly_eval` and
`numeric.poly_grad` when the leading axes are not allowed to be singleton.

In practice, we don't need this flexibility — coefficients never vary per point
— and it is sufficient to treat the leading axes of coordinates and
coefficients similar to an outer product: the first `coords.ndim-1` axes of the
result of `numeric.poly_eval` originate from `coords` and the next
`coeffs.ndim-coords.shape[-1]` axes originate from `coeffs`. This patch
implements this approach.