cache evaluable.as_canonical_length
In a follow-up commit the element loop of `Sample._eval` is moved into an
`evaluable.Array`. The length of this array typically is the sum of non-zeros
over all elements, which is defined using another element loop arrau. The shape
of an `evaluable.Array` is preprocessed using `evaluable.as_canonical_length`,
which evaluates the length if it is constant. This is potentially an expensive
operation. To prevent repeated evaluation, this commit adds an lru cache to
`evaluable.as_canonical_length`.