fix nutilshash for nested dataclasses
This patch fixes the implementation of `types.nutils_hash` for dataclasses with
dataclass attributes. Due to the use of `dataclasses.asdict` these objects were
formerly hashed as dictionaries. One particularly problematic situation was an
attribute of annotated type `typing.Union[mydataclass1, mydataclass2]`, which
resulted in the same hash regardless of the chosen type if the two classes had
matching (or no) attributes. Unfortunately the fix is backwards incompatible,
potentially resulting in cache misses.