fix incorrect usage _WithoutPoints
The `function._WithoutPoints` class is a wrapper that prevents adding points
axes to the lowered `evaluable.Array`. This is mostly useful for evaluables
that accept only scalar indices, such as `evaluable.Elemwise`. It was not
intended to be used to wrap `function.Array` instances that are publicly
exposed, like `function.Array.shape` or `function.transforms_index`, as this
breaks using the index in expressions that do require a points axis:

    topo, geom = mesh.rectilinear([2])
    altgeom = function.rootcoords(1) + topo.f_index

The `rootcoords` can only be lowered with points axis, but `topo.f_index`,
wrapped in `function._WithoutPoints`, always lowers to an array without points
axis.

This commit removes incorrect usage of `function.Array._withoutpoints` and
prevents further mistakes by not subclassing `function.Array`.