fix desparsify Product, Determinant, Sign
Since `Evaluable._desparsify` may return duplicate indices, either in a single
array or in multiple blocks, non-linear operations that preserve zeros cannot
be applied directly to the desparsified values. The implementations of
`Product`, `Determinant` and `Sign` currently produce incorrect results given
duplicate indices, as is demonstrated by the following:

    from nutils.evaluable import Product, InsertAxis, Inflate, Sparse
    import numpy
    a = Product(InsertAxis(Inflate([1,2], [0,0], 1), 2))
    assert isinstance(a._axes[0], Sparse)
    actual = a.eval()
    desired = numpy.zeros_like(actual)
    for ind, val in a._desparsify(0):
      numpy.add.at(desired, ind.eval(), val.eval())
      numpy.testing.assert_equal(desired, actual)

This commit removes the implementation of desparsify for the three affected
evaluables.