combine LoopConcatenate instances
If we evaluate a function that contains several `LoopConcatenate` instances
with shared dependencies, the shared dependencies would be evaluated again and
again in each loop.

This commit adds the internal `Evaluable._combine_loop_concatenates` method
that merges all `LoopConcatenate` instances with the same index argument and
length into a single `LoopConcatenateCombined`.