add {function,evaluable}.Array.as_evaluable_array
In a follow-up commit the `sample.Integral` will be replaced by a
self-contained instance of `function.Array` that is lowerable without arguments
to an `evaluable.Array`. Any function that consumes an `evaluable.Array`,
possibly enclosed in a `sample.Integral`, should in the future work with
anything that can be converted to an `evaluable.Array`. This commit adds the
`as_evaluable_array()` method to `function.Array` and `evaluable.Array` for
automatically converting into an `evaluable.Array`, regardless the source
value.