remove site searches in util.loadlib
For a library that does not reside in the system path it may not be sufficient
to load it by its absolute path, as any dependencies that are loaded indirectly
will not be resolved. Instead we must set this path a priori, and can thus
remove any fall-back search logic from util.loadlib.