don't retrieve geom in jump expr. if unnecessary
The normal jump, e.g. `[u]_i` as shorthand for `[u] n_i`, existed briefly, but
has been deprecated for a while now. When parsing the jump expresion, we
retrieve the default geometry even if we are not parsing a jump with normal.
This is problematic if we are using a different variable for the geometry, say
`t`, than what the `function.Namespace` assumes to be the default geometry,
normally `x`. This patch defers retrieving the geometry.