fix Inflate explicit infl. check with not 1d dofs
The `Inflate` evaluable warns about expensive evaluation if the dofmap is
smaller than the length of the inflated axis. The check incorrectly assumes the
dofmap is 1d, while `Inflate` actually accepts a dofmap of any dimension. This
patch fixes the problem by looking at the size of the dofmap instead of the
length of the first axis and adds a unit test with a scalar dofmap.