replace `d(u, x_i)` with `d(u, x)_i`
The expression `d(u, x_i)` should be interpreted as the gradient of `u` w.r.t.
geometry `x`, and the resulting axis is identified with index `i`. This breaks
the assumption that functions in the namespace are treated pointwise, except
for the consumed axes. To fix this problem, this patch redefines the gradient
operator `d(...)` such that all axes of the second argument, the geometry, are
consumed and the same amount of axes, with the same shape, are generated:
`d:j(u, x_j)_i` consumes `j` and generates `i`. For brevity, this patch also
supports autoconsuming all axes of arguments parseable without indices: `d(u,
x)_i`.

Fixes #614