add todims to Transforms sequence
The `Transforms` sequence consists of a sequence of transform chains that all
have the same from and to dimension. The from dimension is available as
attribute `Transforms.fromdims`, but `Transforms.todims` is not. In a follow-up
patch the reference coordinate system for taking gradients is changed from tip
to root. This requires defining the dimension of the root, which is used as the
length of the axes of the root derivative target. We have two options to define
this: either as an `evaluable.Array` that evaluates to the to dimension of the
current transform chain, or we add the to dimension as an attribute to the
transforms sequence. This patch does the latter.