add spaces to function.Array
Currently the `ProductTopology` relies on the bifurcate family of classes and
functions to join and separate transform chains originating from parent
topologies. This family will be replaced by the concept of spaces.

A 'space' is a Euclidean space of a certain dimension. It is identified by a
name (`str`). Every `Topology` and `Sample` has an ordered list of unique
spaces. Every element in the topology or sample is identified by a tuple of
transform chains, one for each space. The tuples of transform chains are not
necessarily maintained in a `Transforms` sequence, in part because these
sequences only cover single transform chains. `function.Array` instances can be
defined on a subset (including all or none) of spaces of a topology or sample.
Functions are assumed to be constant in spaces other than the spaces on which
they are defined, hence functions can be evaluated on a superset of spaces.

This patch adds spaces to `function.Array`.