add function._root_derivative_target
Currently gradients, normals and jacobians are computed with the local or tip
coordinate system as a reference. The root coordinate systems has one advantage
over the tip coordinates: the reference is independend of the current set of
transform chains. In a series of patches the root coordinate system will be
made the reference.

This patch introduces the `function._root_derivative_target`, which will be
used to compute derivative to the root coordinate system of some space, similar
to the `evaluable.LocalCoords`.