simplify finitecell manifold unittest
Before introducing tensorial topologies, the reference coordinate system for
computing gradients must change from tip, e.g. the elements of a `Sample`, to
roots. It is yet uncertain whether is it possible to compute gradients on
manifolds to geometries with a dimension equal to the manifold, hence smaller
than the dimension of the root. The `cutdomain` finite cell test does exactly
this. Anticipating the change of the reference coordinate system, this patch
simplifies the test to full domains, not manifolds.