add evaluable.WithDerivative
`Array` instances define derivatives to `DerivativeTargetBase`, a subclass of
`Array`. Currently there are two kinds of targets: `Argument`, a rather
straightforward target with the identity as self-derivative, and `LocalCoords`,
which is used for derivatives to a coordinate system, described by
transformations (`nutils.transform.Transform`). Derivatives to the latter end
up in `ApplyTransforms`, at which point the derivative is defined as the linear
part of the transform of `ApplyTransforms`.

In a follow-up patch the reference coordinate system for computing gradients,
normals et cetera, will be changed from the tip or the current transform chain
(currently called the local coordinate system) to the root.

Derivatives to the root require a virtual derivative target, for which the
`IdentifierDerivativeTarget` will be used - similar to `LocalCoords`, but
generic - and a special case implementation for the derivative of some `Array`
to this target. This patch adds a generic `Array` wrapper for defining a
special case derivative for an `Array`.