impl root, tip gradient for function.RootCoords
Currently gradients, normals and jacobians are computed with the local or tip
coordinate system as a reference. The root coordinate systems has one advantage
over the tip coordinates: the reference is independend of the current set of
transform chains. In a series of patches the root coordinate system will be
made the reference.

This patch implements the derivative of `function.RootCoords` to the root and
tip coordinate system of a given space.