add EvaluableTransformChain
The `evaluable` module has a `TransformChain` base class with the following
problem: The to dimension is possibly undefined and the from dimension is
always undefined. This has the side effect that we need to provide both the to
and the from dimension if we want to compute the linear transformation matrix.

This patch adds a fresh `EvaluableTransformChain` to the `transform` module,
with mandatory to and from dimensions and convenience attributes for computing
linears, normals, et cetera. The old `evaluable.TransformChain` class and
subclasses will be removed in a follow-up commit. To ease the transition,
several functions in the `evaluable` modules are adapted such that they work
with both the old and the new evaluable transform chains.