fix dtype of Normal.evalf for 1d argument
The `evaluable.Normal` class has a fast implementation of `evalf` for
one-dimensional geometries. This implementation, however, returns an array with
the same dtype as the argument of `evalf`, the local gradient of the geometry,
which could be different from the dtype announced by `evaluable.Normal`:
`float`. This patch fixes the problem by adding an explicit cast to a `float`.