simplify Normal of 1D geometry to Sign
The normal at the edge of a line is the sign of the local gradient of the
geometry. The `evaluable.Normal` has a fast implementation of `evalf` for this
situation. This patch replaces the fast evalf with a simplification of the same
kind in case the dimension of the geometry is constant one and adds unittests.