add loop_index for loop_sum, loop_concatenate
The `evaluable.InRange` evaluable checks the bounds of its argument at runtime.
If `InRange` is applied to the index of a `LoopSum` or `LoopConcatenate`, we
could simplify the `InRange` away, provided that we can get hold of the length
of the loop. This patch introduces `evaluable.loop_index` with an explicit
reference to the length of the loop and mandates that the loop index of
`evaluable.LoopSum` and `evaluable.LoopConcatenate` is of this class.