drop offset from evaluable.InRange
The `InRange` evaluable takes `length`, `offset` and `index` as arguments. In
`evalf` the `index` is asserted to be within 0 and `length` after which the
`offset` is added to `index` and returned. This patch removes the `offset`
argument in favor of an explicit `Add(InRange(...), offset)` and reorders the
remaining to `index` and `length`.