fix MKL tests in Github Actions
The most recent version of MKL on PyPI ships only `libmkl_rt.so.1`, not
`libmkl_rt.so`. In Nutils we load the latter. This causes two problems: in
Github Actions we cannot find the location of the library directory where MKL
is being installed, because we explicitly look for `libmkl_rt.so` and Nutils
can't load the library because we explicitly load the missing unversioned
library. This patch fixes both problems by searching for the versioned library
and adding a symlink to the unversioned library.