fix invalid length error with meshio.CellBlock
In commit [60d11b70] of meshio the length of `CellBlock`, a subclass of
`namedtuple`, is redefined to the length of the `data` attribute. This breaks
`namedtuple._replace`, where the length of the newly created `CellBlock` is
verified. This patch circumvents the problem by using the `CellBlock`
constructor directly instead of `namedtuple._replace`.

[60d11b70]: https://github.com/nschloe/meshio/commit/60d11b705c9b9074284feefdb7475db0410953e9