disallow Sum of boolean array
This patch disallows a boolean array as input. The main motivation is that a
boolean array does not have `_intbounds`, and this causes the `_intbounds` of
`Sum` to be infinite with the current implementation. The cast function `Int`
already defines `_intbounds` for boolean input, hence `Sum(Int(mask))` has
proper `_intbounds`.