add function.Custom
Since `evaluable` has been split off of `function` it is more difficult for end
users to implement their own `function.Array` as they also have to implement an
`evaluable.Array`. To lessen the burden, this commit introduces
`function.Custom`: a base class that automatically lowers to an
`evaluable.Array` with `evalf` and `derivative` taken from the implementation
of `function.Custom`.