add points_shape lower argument
Currently the shape of the leading points axes that should be present in the
lowered `evaluable.Array` is distilled from the `coordinates` argument. This
commit generalizes the concept of leading points axis by introducing a
dedicated argument: `points_shape`. With this it is possible to request points
axes during lowering without defining coordinates. If coordinates are present,
the shape of the coordinates must match the shape of the points axes.