add types.hashable_function
Plain functions are not Nutils hashable, because there is no reliable
information to generate a hash from: the name of the function cannot be used
because the behavior might be non-deterministic or can simply be changed
between different runs, and the bytecode only describes a single function, not
the functions that are being called or variables that are being accessed from
outer scopes, if any. I hear you say: 'but the dependent functions can be
distilled from the bytecode,' and: 'the closure variables are stored in the
function object.' Well, I wish you good luck with implementing a solid hash in
a Tuesday evening. This leaves us with the possibility to explicitly add a
Nutils hash to functions that need one. This commit introduces the decorator
`types.hashable_function` for this purpose.