add function._Unlower
Once lowered, a `evaluable.Array` cannot be converted back into the original
`function.Array`. It is, however, possible to wrap the `evaluable.Array` in
some other `function.Array` which can be lowered to the former provided that
the original lower arguments are used. This commit introduces
`function._Unlower` for this purpose.

`function._Unlower` will be used in a follow-up commit for the evaluation of a
partial derivative of some `evaluable.Array`, which is defined in terms of
`function.Array`.