fix ufunc dtype incompatibility
Since NumPy version 1.21 the [`dtype` argument of ufuncs must be a basic dtype
sans byteorder or other metadata][1], i.e. `numpy.float64`. In `sparse.blocks`
we explicitly define the dtype in which the computation of a ufunc should be
perfomed using a fat dtype. This commit slims this dtype down to basic.

[1]: https://github.com/numpy/numpy/blob/v1.21.0/doc/source/release/1.21.0-notes.rst#changes-to-dtype-and-signature-arguments-in-ufuncs