remove Lowerable as Array base class
The `Lowerable` class is a protocol that describes an interface for lowering an
instance to an `evaluable.Array`. The `function.Array` implements this
interface. Currently `function.Array` explicitly derives from `Lowerable`. This
is however not a requirement and the explicit inheritance can be removed,
thereby loosing only the ability to do an `isinstance(..., Lowerable)` on
instances of `function.Array`.

In a follow-up patch the `Lowerable` interface is extended with a `spaces`
property. In `function.Array` this will be implemented using an attribute. This
is, however, incompatible with the `spaces` property as defined by the
`Lowerable` base class. To circumvent this problem, this patch removes
`Lowerable` as base cass from `function.Array`.