add new Topology base class sans transform chains
This patch introduces a new `Topology` base class without the notion of
transform chains. The new base class has the exact same api as the old class,
for now. The old class is renamed to `TransformChainsTopology` and all methods
that are independent of the `transforms` and `opposites` attributes have been
moved into the new `Topology` base class, without any modification. All other
methods remain in the old class and are not implemented in the new class. The
new base class is annotated according to [PEP484].

[PEP484]: https://www.python.org/dev/peps/pep-0484/