fix periodic mesh.line
If a periodic `mesh.line` is created, the basis that is used to build the
geometry should be nonperiodic. Since the topology is a `StructuredTopology`,
the basis accepts a *list* of periodic dimension indices for parameter
`periodic`, not a boolean. This patch fixes the mistake by passing an empty
list to `topo.basis` instead of `False`.