fix doctest float neighborhood output checker
The doctest output checker for floats with `±` ranges splits the output in
chunks, alternating between plain strings and floats with neighborhoods. The
regular expression that performs the split matched strings of the form `1.±...`
and `.1±...`, but not `1.2±...`. Outputs that contained floats of last form
passed the check in all current cases by sheer luck, because the leading digits
were considered to be part of the nonfloat chunks and the actual and desired
leading digits matched. This patch fixes the problem nevertheless by also
accepting floats of the form `1.2±...`.