force dtype of Reciprocal to float
`numpy.reciprocal` returns an array with the same dtype as its argument.
`evaluable.Reciprocal` has `numpy.reciprocal` at its `evalf`, hence copies this
behavior. The `evaluable.Reciprocal`, however, is used in the `evaluable`
module with the assumption that the result is not truncated. This patch forces
the dtype of the result to `float`.