fix LoopSum.asarray assertion
In `LoopSum.asarray` the axes that are invariant to the loop index are shuffled
to the start to minimize the number of ravels. After the shuffle and ravels all
but the last axes are asserted to be invariant to the loop index. The check,
however, incorrectly checks for constness, which is more strict than being
invariant to the loop index. This patch fixes the assertion and removes an
`int` check, which is always false since the shapes are always of type
`evaluable.Array`.