prioritize LoopSum over Multiply
If we have `Multiply <- LoopSum <- Inflate` function tree, then `Multiply`
prohibits the sparsification of `LoopSum`. This situation arise a.o. from
nested integrals, where the `Multiply` is part of the inner product of inner
integral with integration weights of the outer integral. This patch fixes the
problem by prioritizing `LoopSum` over `Multiply`.