impl Ravel._inflations
Because `Ravel` has a higher priority than `Inflate`, it seems unnecessary to
implement `Ravel._inflations`. However, if the `Ravel` is blocked by some other
array, for instance a `LoopConcatenate`, then inflated axes cannot move down
the function tree and we miss potential simplifications. This patch implements
`Ravel._inflations` to make this kind of simplifications possible.

In addition this patch defines `Ravel._inflations` for the raveled axis if one
or both of the axes to be raveled is inflated. This makes it possible to
simplify `Take(Ravel(Inflate(...)))` to `Inflate(Take(...))`. The former arises
from slicing a basis on a product topology, which is a raveled outer product of
bases.