add RavelIndex
This patch adds a specific evaluable for the operation ia*nb+ib, that computes
an index for a raveled axis based on two indices for the original, unraveled
axes. While the operation itself can (and used to) be formed using the existing
multiplication and addition operations, the new operation allows for specific
simplifications to be performed that rely on this particular structure, which
would otherwise be difficult to recognize.