simplify Take of covered Inflate
Currently, if the argument of `Take` is inflated on the last axis, but the
`_take` simplification doesn't reach an `Inflate`, for instance because the
`Inflate` is covered by a `Ravel`, then the `Inflate` and `Take` will not be
swapped. This patch adds support for this kind of simplification by using the
`_inflations` property of the argument of `Take`.