correctly close teelog before htmllog
This patch replaces set(TeeLog(console, html)) by set(console), add(html). The
issue with the former is that in treelog 2.0, upon leaving, set reinserts the
teelog, which was not itself entered and therefore will not reinsert console
and html until it is garbage collected. If this happens after html is closed
via its context then this will fail for being currently in use. The new setup
correctly reinserts html before closing it.