prevent parsing func calls as old-style operators
The old-style operators `d`, `n` and `J` - not to be confused with the
functions with the same name, included by default in the namespace - are
identified as such during tokenization. The tokenizer incorrectly matches an
operator when followed by a left parenthesis symbol. As a consequence it is
impossible to parse a function-style operator under certain conditions. For
example `d:i(u, x_i)_j` is currently tokenized as `jacobian of i, '(', ...`
instead of `variable 'd', consumes 'i', '(', ...`. This patch fixes the problem
by adding a look-ahead test for left parentheses after a potential match of an
old-style operator.