try numpy.asarray in function.Array.cast
If we call `function.Array.cast` with a list, the list is passed to
`function.take` with each element of the (outer) list cast to a
`function.Array`. If the list is a (possibly nested) list of numbers this is
unlikely the desired outcome. For example

    Array.cast([0, 1, 2])

is equivalent to

    stack([_Constant(0), _Constant(1), _Constant(2)])

Instead, this patch tries to cast the value to a `numpy.ndarray` first and, if
that works with numeric dtype, proceeds with a `function.Constant`.