support tri, hull for single point products
This patch adds the tri and hull attributes to multiplication samples of which
one of the factors is a single point sample, by simply passing on the tri or
hull of the other factor. This adds support for sampling boundaries of extruded
topologies that are not themselves products of lines.