add function.dotarg
In follow-up commits a new expression parser and namespace will be introduced.
The new expression parser cannot infer lengths of indices. For this reason the
new parser does not support the `'?'+name` syntax for arguments. This patch
adds an alternative in the form of `function.dotarg`: the inner product of an
array, typically a basis, with a to be created argument.