change derivative of min, max
This patch changes the derivative of min and max from an implementation based
on less to one based on sign. Even though this changes the derivative only at
the break-even point where it is not defined, this fixing a unit test issue in
test_function.check where the reference value is computed via symmetric finite
difference.