replace sample._AtSample by _Concat&_ReorderPoints
This patch changes the implementation of sample.__matmat__ from producing a
sample._AtSample, which lowered to a loopsum of an inflation, to a combination
of _ConcatenatePoints and _ReorderPoints, which lowers to a loopconcatenation.
Besides being closer to the form in which the evaluation will eventually take
place, the new construction also avoids summation which is not well defined in
case of boolean arrays. Furthermore, the new structure allows for a
simplification or _DefaultIndex.__matmul__, which can skip _ReorderPoints.