verify jacobian dimension if specified
During lowering of the `function._Jacobian` the dimensions of the tip
coordinates per space are known. Since the dimension of the tip coordinates is
not used before lowering, it is allowed to omit the dimension of the tip
coordinates when calling `function.jacobian`.

In the future we may incorporate units into the `function.Array`. The unit of
the jacobian is the n-th power of the unit of the geometry, with `n` the
dimension of the tip coordinates, hence if we want to define a unit for the
jacobian, we must know this dimension.

Anticipating on the introduction of units, this patch adds a check of the tip
dimension during lowering of `function._Jacobian` if this dimension is
specified (e.g. via `function.jacobian`).