reimpl f.Array.derivative using f.Array.arguments
Taking the derivative of a `function.Array` to an argument requires knowing the
shape of the argument. When `function.Array.derivative` is called with a `str`,
the shape will be obtained by lowering the array to an `evaluable.Array` and
looking up the argument in the `arguments` property. This is an expensive
operation and only works if all spaces of the `function.Array` are bound. This
patch reimplements `function.Array.derivative` using the new
`function.Array.arguments` attribute, which is inexpensive and is always
available.