reimpl f.Array.argshapes using f.Array.arguments
In absence of the `function.Array.arguments` attribue, the
`function.Array.argshapes` property was implemented by lowering to a
`evaluable.Array` an collecting all `evaluable.Arguments`. This is an expensive
operation and only works if all spaces of the `function.Array` are bound. This
patch reimplements `function.Array.argshapes` using the new
`function.Array.arguments` attribute, which is inexpensive and is always
available.