fix axis sample._ReorderPoints
`Sample._ReorderPoints` is applied after `Sample._ConcatenatePoints` to reorder
the points according to the indices of the sample. The latter prepends an axis
to its argument, hence the first axis of the `function.Array` is the points
axis of the sample in question. When lowering the `Sample._ConcatenatePoints`
to an `evaluable.Array`, this axis comes *after* all points axes introduced by
outer integrals and samples during lowering. The implementation of
`Sample._ReorderPoints.lower`, however, incorrectly assumed the axis to reorder
was the first axis. This patch fixes the implementation.