make function.Basis.lower test less strict
The basis test suite automatically tests the implementation of `lower` against
polyvals of the `checkcoeffs` attribute of the test class. Rounding errors in
the former and the latter may cause the test to fail. All current bases have
the same implementation for `lower` as is used in the test, hence testing for
equality suffices. In a follow-up commit a basis with an entirely different
implemenation of `lower` is introduced, which causes this test to fail. This
patch resolves the problem by rounding the results to ten digits before
checking equality.