fix reordering points in func @ sample
In commit 6266ca9a the `sample._AtSample` function has been split into a
concatenate and reorder. The latter is omitted when the indices are trivial:
`[0,1,2,...]`. The reorder function has been implemented using `Take` instead
of `Inflate`, because the former is slightly cheaper than the latter. However,
we forget about inverting the indices, which is required for `Take` and much
more expensive than `Inflate`. This patch replaces `Take` with `Inflate` and
adds and amends several unit tests.